+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 - **Change:** Increase minimum XenForo Resource Manager version requirement to 2.1.4
 - **Change:** Use consistent listener execution order (#2)
 - **Fix:** XenForo Resource Manager 2.2 compatibility (#1)
+- **Fix:** Changing resource type does not set update user data (#3)
 
 ## 1.0.0 (`1000070`)